+++ gluecodium/src/main/java/com/here/gluecodium/cli/OptionReader.kt
@@ -66,6 +66,21 @@ object OptionReader {
                 "Java @Nullable annotation full name including package",
             )
             addOption(
+                "androidoptinannotation",
+                true,
+                "Android (Kotlin) @OptIn annotation full name including package",
+            )
+            addOption(
+                "androidrequiresoptinannotation",
+                true,
+                "Android (Kotlin) @RequiresOptIn annotation full name including package",
+            )
+            addOption(
+                "androidinternalapiannotationname",
+                true,
+                "The name for Android (Kotlin) internal API annotation (defaults to 'GluecodiumInternalApi')",
+            )
+            addOption(
                 "intpackage",
                 "java-internal-package",
                 true,
@@ -240,6 +255,9 @@ object OptionReader {
 
         generatorOptions.kotlinPackages = getStringValue("kotlinpackage")?.split(".") ?: emptyList()
         generatorOptions.kotlinInternalPackages = getStringValue("kotlinintpackage")?.split(".") ?: emptyList()
+        generatorOptions.androidOptInAnnotation = getStringValue("androidoptinannotation")?.split(".")
+        generatorOptions.androidRequiresOptInAnnotation = getStringValue("androidrequiresoptinannotation")?.split(".")
+        generatorOptions.androidInternalApiAnnotationName = getStringValue("androidinternalapiannotationname") ?: "GluecodiumInternalApi"
 
         generatorOptions.cppRootNamespace = getStringValue("cppnamespace")?.split(".") ?: emptyList()
         generatorOptions.cppInternalNamespace = getStringValue("intnamespace")?.split(".") ?: emptyList()