+++ gluecodium/src/main/java/com/here/gluecodium/generator/kotlin/KotlinGenerator.kt
@@ -59,6 +59,9 @@ internal class KotlinGenerator : Generator {
     private lateinit var kotlinNameRules: KotlinNameRules
     private lateinit var activeTags: Set<String>
     private lateinit var werror: Set<String>
+    private var internalApiAnnotationName: String? = null
+    private var requireOptInAnnotation: List<String>? = null
+    private var optInAnnotation: List<String>? = null
 
     override val shortName = GENERATOR_NAME
 
@@ -72,6 +75,22 @@ internal class KotlinGenerator : Generator {
         kotlinNameRules = KotlinNameRules(nameRuleSetFromConfig(options.kotlinNameRules))
         activeTags = options.tags
         werror = options.werror
+        requireOptInAnnotation = options.androidRequiresOptInAnnotation
+        internalApiAnnotationName = options.androidInternalApiAnnotationName
+        optInAnnotation = options.androidOptInAnnotation
+
+        if (!preservesOptInAnnotationConsistency()) {
+            throw GluecodiumExecutionException(
+                "Kotlin generator requires either both annotations specified ('RequiresOptIn' and 'OptIn') or none of them!",
+            )
+        }
+    }
+
+    private fun preservesOptInAnnotationConsistency(): Boolean {
+        val bothSpecified = requireOptInAnnotation != null && optInAnnotation != null
+        val noneSpecified = requireOptInAnnotation == null && optInAnnotation == null
+
+        return bothSpecified || noneSpecified
     }
 
     override fun generate(limeModel: LimeModel): List<GeneratedFile> {
@@ -108,11 +127,13 @@ internal class KotlinGenerator : Generator {
 
         val visibilityResolver = KotlinVisibilityResolver(limeModel.referenceMap)
 
+        val internalApiAnnotation = if (requireOptInAnnotation != null) internalApiAnnotationName!! else null
         val importResolver =
             KotlinImportResolver(
                 limeReferenceMap = limeModel.referenceMap,
                 nameResolver = nameResolver,
                 internalPackage = internalPackageList,
+                internalApiAnnotation = internalApiAnnotation,
             )
 
         val importCollector =
@@ -122,7 +143,16 @@ internal class KotlinGenerator : Generator {
 
         val resultFiles =
             kotlinFilteredModel.topElements
-                .flatMap { generateKotlinFiles(it, nameResolver, visibilityResolver, importResolver, importCollector) }
+                .flatMap {
+                    generateKotlinFiles(
+                        it,
+                        nameResolver,
+                        visibilityResolver,
+                        importResolver,
+                        importCollector,
+                        internalApiAnnotation,
+                    )
+                }
                 .toMutableList()
 
         val nativeBasePath = (listOf(GENERATOR_NAME) + internalPackageList).joinToString("/")
@@ -144,6 +174,21 @@ internal class KotlinGenerator : Generator {
                 "$nativeBasePath/time/Duration.kt",
                 GeneratedFile.SourceSet.COMMON,
             )
+
+        if (requireOptInAnnotation != null) {
+            val modelData =
+                mapOf(
+                    "internalPackageList" to internalPackageList,
+                    "requireOptInAnnotation" to requireOptInAnnotation,
+                    "internalApiAnnotationName" to internalApiAnnotationName,
+                )
+            resultFiles +=
+                GeneratedFile(
+                    TemplateEngine.render("kotlin/KotlinInternalApiAnnotation", modelData),
+                    "$nativeBasePath/${internalApiAnnotationName!!}.kt",
+                    GeneratedFile.SourceSet.COMMON,
+                )
+        }
 
         val descendantInterfaces = LimeTypeHelper.collectDescendantInterfaces(jniFilteredModel.topElements)
         val jniTemplates =
@@ -191,6 +236,7 @@ internal class KotlinGenerator : Generator {
         visibilityResolver: KotlinVisibilityResolver,
         importResolver: KotlinImportResolver,
         importCollector: KotlinImportCollector,
+        internalApiAnnotation: String?,
     ): List<GeneratedFile> {
         if (limeElement.external?.kotlin?.get(NAME_NAME) != null &&
             limeElement.external?.kotlin?.get(CONVERTER_NAME) == null
@@ -202,6 +248,9 @@ internal class KotlinGenerator : Generator {
         val packages = (basePackages + limeElement.path.head).map { KotlinNameResolver.normalizePackageName(it) }
         val imports = importCollector.collectImports(limeElement).filterNot { KotlinNameRules.getPackageFromImportString(it) == packages }
         val optimizedLists = OptimizedListsCollector().getAllOptimizedLists(limeElement)
+        val optInAnnotationString = optInAnnotation?.joinToString(".")
+        val internalApiAnnotationClassPath =
+            if (optInAnnotationString != null) internalPackageList.joinToString(".") + "." + internalApiAnnotation else null
 
         val templateData =
             mutableMapOf(
@@ -211,6 +260,9 @@ internal class KotlinGenerator : Generator {
                 "package" to packages,
                 "imports" to imports.distinct().sorted(),
                 "optimizedLists" to optimizedLists,
+                "optInAnnotationString" to optInAnnotationString,
+                "internalApiAnnotation" to internalApiAnnotation,
+                "internalApiAnnotationClassPath" to internalApiAnnotationClassPath,
             )
 
         val nameResolvers = mapOf("" to nameResolver, "visibility" to visibilityResolver)