+++ gluecodium/src/test/java/com/here/gluecodium/cli/OptionReaderTest.kt
@@ -197,6 +197,42 @@ class OptionReaderTest {
     }
 
     @Test
+    fun androidrequiresoptinannotationOptionIsRecognised() {
+        // Arrange
+        val toRead = prepareToRead("-androidrequiresoptinannotation", TEST_ANDROID_REQUIRES_OPT_IN_ANNOTATION)
+
+        // Act
+        val options = OptionReader.read(toRead)
+
+        // Assert
+        assertEquals(listOf("androidx", "annotation", "RequiresOptIn"), options!!.second.androidRequiresOptInAnnotation)
+    }
+
+    @Test
+    fun androidoptinannotationOptionIsRecognised() {
+        // Arrange
+        val toRead = prepareToRead("-androidoptinannotation", TEST_ANDROID_OPT_IN_ANNOTATION)
+
+        // Act
+        val options = OptionReader.read(toRead)
+
+        // Assert
+        assertEquals(listOf("androidx", "annotation", "OptIn"), options!!.second.androidOptInAnnotation)
+    }
+
+    @Test
+    fun androidinternalapiannotationnameOptionIsRecognised() {
+        // Arrange
+        val toRead = prepareToRead("-androidinternalapiannotationname", TEST_ANDROID_INTERNAL_API_ANNOTATION)
+
+        // Act
+        val options = OptionReader.read(toRead)
+
+        // Assert
+        assertEquals(TEST_ANDROID_INTERNAL_API_ANNOTATION, options!!.second.androidInternalApiAnnotationName)
+    }
+
+    @Test
     fun docsplaceholderslistMissingFile() {
         // Arrange, Act
         val path = "someFileThatDoesNotExists.properties"
@@ -234,5 +270,8 @@ class OptionReaderTest {
         private const val TEST_GENERATORS = "java,cpp"
         private const val TEST_JAVA_PACKAGE_LIST = "some_package"
         private const val TEST_KOTLIN_PACKAGE_LIST = "another_package"
+        private const val TEST_ANDROID_OPT_IN_ANNOTATION = "androidx.annotation.OptIn"
+        private const val TEST_ANDROID_REQUIRES_OPT_IN_ANNOTATION = "androidx.annotation.RequiresOptIn"
+        private const val TEST_ANDROID_INTERNAL_API_ANNOTATION = "MyLibraryInternalApi"
     }
 }