+++ gluecodium/src/test/resources/smoke/annotations/output/android-kotlin/com/example/smoke/Annotations.kt
@@ -1,0 +1,42 @@
+/*
+
+ *
+ */
+
+@file:JvmName("AnnotationsExtensions")
+@file:androidx.annotation.OptIn(com.example.MySmokeTestsInternalApi::class)
+
+package com.example.smoke
+
+import com.example.MySmokeTestsInternalApi
+import com.example.NativeBase
+
+@MySmokeTestsInternalApi
+internal class Annotations : NativeBase {
+
+
+
+    /**
+     * For internal use only.
+     * @suppress
+     * @param nativeHandle The handle to resources on C++ side.
+     * @param tag Tag used by callers to avoid overload resolution problems.
+     */
+    protected constructor(nativeHandle: Long, @Suppress("UNUSED_PARAMETER") tag: Any?)
+        : super(nativeHandle, { disposeNativeHandle(it) }) {}
+
+
+
+
+
+    external fun testOptional(self: Annotations) : Boolean?
+
+
+
+
+    companion object {
+        @JvmStatic private external fun disposeNativeHandle(nativeHandle: Long)
+    }
+}
+
+