+++ gluecodium/src/test/resources/smoke/annotations/output/android-kotlin/com/example/smoke/OuterInternalEnum.kt
@@ -1,0 +1,17 @@
+/*
+
+ *
+ */
+
+@file:JvmName("OuterInternalEnumExtensions")
+@file:androidx.annotation.OptIn(com.example.MySmokeTestsInternalApi::class)
+
+package com.example.smoke
+
+import com.example.MySmokeTestsInternalApi
+
+@MySmokeTestsInternalApi
+internal enum class OuterInternalEnum(@JvmField internal val value: Int) {
+    FIRST(1),
+    SECOND(2);
+}