+++ gluecodium/src/test/resources/smoke/annotations/output/android-kotlin/com/example/smoke/OuterInternalException.kt
@@ -1,0 +1,16 @@
+/*
+
+ *
+ */
+
+@file:JvmName("OuterInternalExtensions")
+@file:androidx.annotation.OptIn(com.example.MySmokeTestsInternalApi::class)
+
+package com.example.smoke
+
+import com.example.MySmokeTestsInternalApi
+
+@MySmokeTestsInternalApi
+internal class OuterInternalException(@JvmField val error: OuterInternalEnum) : Exception(error.toString())
+
+