+++ gluecodium/src/test/resources/smoke/annotations/output/android-kotlin/com/example/smoke/OuterInternalInterface.kt
@@ -1,0 +1,22 @@
+/*
+
+ *
+ */
+
+@file:JvmName("OuterInternalInterfaceExtensions")
+@file:androidx.annotation.OptIn(com.example.MySmokeTestsInternalApi::class)
+
+package com.example.smoke
+
+import com.example.MySmokeTestsInternalApi
+
+@MySmokeTestsInternalApi
+internal fun interface OuterInternalInterface {
+
+
+
+    fun someFunction() : Int
+
+
+}
+