+++ gluecodium/src/test/resources/smoke/annotations/output/android-kotlin/com/example/smoke/OuterInternalLambda.kt
@@ -1,0 +1,19 @@
+/*
+
+ *
+ */
+
+@file:JvmName("OuterInternalLambdaExtensions")
+@file:androidx.annotation.OptIn(com.example.MySmokeTestsInternalApi::class)
+
+package com.example.smoke
+
+import com.example.MySmokeTestsInternalApi
+
+@MySmokeTestsInternalApi
+internal fun interface OuterInternalLambda {
+
+
+    fun apply(p0: String) : Int
+}
+